Package model.tariff renamed
