+++ src/main/java/ru/itpark/model/tariff/Base.java
@@ -1,4 +1,4 @@
-package ru.itpark.model.tariffs;
+package ru.itpark.model.tariff;
 
 
 public class Base<T> {
